Validate repository URLs in darc subscription commands
