+++ src/Microsoft.DotNet.Darc/Darc/Models/PopUps/UpdateSubscriptionPopUp.cs
@@ -1,26 +1,20 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.DotNet.DarcLib;
 using Microsoft.DotNet.Maestro.Client.Models;
 using Microsoft.Extensions.Logging;
-using YamlDotNet.Serialization;
 
 #nullable enable
 namespace Microsoft.DotNet.Darc.Models.PopUps;
 
-public class UpdateSubscriptionPopUp : SubscriptionPopUp
+internal class UpdateSubscriptionPopUp : SubscriptionPopUp<SubscriptionUpdateData>
 {
-    public const string EnabledElement = "Enabled";
-
     private readonly ILogger _logger;
-    private readonly SubscriptionUpdateData _yamlData;
 
-    public bool Enabled => bool.Parse(_yamlData.Enabled);
+    public bool Enabled => bool.Parse(_data.Enabled);
 
     private UpdateSubscriptionPopUp(
         string path,
@@ -30,9 +24,10 @@ public class UpdateSubscriptionPopUp : SubscriptionPopUp
         Subscription subscription,
         IEnumerable<string> suggestedChannels,
         IEnumerable<string> suggestedRepositories,
+        IEnumerable<string> availableUpdateFrequencies,
         IEnumerable<string> availableMergePolicyHelp,
         SubscriptionUpdateData data)
-        : base(path, forceCreation, suggestedChannels, suggestedRepositories, availableMergePolicyHelp, logger, gitRepoFactory, data,
+        : base(path, forceCreation, suggestedChannels, suggestedRepositories, availableUpdateFrequencies, availableMergePolicyHelp, logger, gitRepoFactory, data,
             header: [
                 new Line($"Use this form to update the values of subscription '{subscription.Id}'.", true),
                 new Line($"Note that if you are setting 'Is batchable' to true you need to remove all Merge Policies.", true),
@@ -47,7 +42,6 @@ public class UpdateSubscriptionPopUp : SubscriptionPopUp
             ])
     {
         _logger = logger;
-        _yamlData = data;
     }
 
     public UpdateSubscriptionPopUp(
@@ -65,7 +59,7 @@ public class UpdateSubscriptionPopUp : SubscriptionPopUp
         string sourceDirectory,
         string targetDirectory,
         List<string> excludedAssets)
-        : this(path, forceCreation, gitRepoFactory, logger, subscription, suggestedChannels, suggestedRepositories, availableMergePolicyHelp,
+        : this(path, forceCreation, gitRepoFactory, logger, subscription, suggestedChannels, suggestedRepositories, availableUpdateFrequencies, availableMergePolicyHelp,
               new SubscriptionUpdateData
               {
                   Id = GetCurrentSettingForDisplay(subscription.Id.ToString(), subscription.Id.ToString(), false),
@@ -86,46 +80,22 @@ public class UpdateSubscriptionPopUp : SubscriptionPopUp
     {
     }
 
-    public override async Task<int> ProcessContents(IList<Line> contents)
+    protected override async Task<int> ParseAndValidateData(SubscriptionUpdateData data)
     {
-        SubscriptionUpdateData outputYamlData;
-
-        try
+        int result = await base.ParseAndValidateData(data);
+        if (result != Constants.SuccessCode)
         {
-            string yamlString = contents.Aggregate(string.Empty, (current, line) => $"{current}{Environment.NewLine}{line.Text}");
-            IDeserializer serializer = new DeserializerBuilder().Build();
-            outputYamlData = serializer.Deserialize<SubscriptionUpdateData>(yamlString);
-        }
-        catch (Exception e)
-        {
-            _logger.LogError(e, "Failed to parse input yaml.  Please see help for correct format.");
-            return Constants.ErrorCode;
+            return result;
         }
 
-        var result = await ParseAndValidateData(outputYamlData);
-
-        if (!bool.TryParse(outputYamlData.Enabled, out bool enabled))
+        if (!bool.TryParse(data.Enabled, out _))
         {
             _logger.LogError("Enabled is not a valid boolean value.");
             return Constants.ErrorCode;
         }
 
-        _yamlData.Enabled = ParseSetting(outputYamlData.Enabled, _yamlData.Enabled, false)!;
+        _data.Enabled = ParseSetting(data.Enabled, _data.Enabled, false)!;
 
-        return result;
-    }
-
-    /// <summary>
-    /// Helper class for YAML encoding/decoding purposes.
-    /// This is used so that we can have friendly alias names for elements.
-    /// </summary>
-    #nullable disable
-    private class SubscriptionUpdateData : SubscriptionData
-    {
-        [YamlMember(ApplyNamingConventions = false)]
-        public string Id { get; set; }
-
-        [YamlMember(Alias = EnabledElement, ApplyNamingConventions = false)]
-        public string Enabled { get; set; }
+        return Constants.SuccessCode;
     }
 }
